Criação de pastas e separação dos arquivos: services, types e connection
