+++ semana19/projeto-labook/src/business/entities/user.ts
@@ -1,0 +1,10 @@
+type AuthenticationData = {
+    id: string
+}
+ 
+type User = {
+    id: string,
+    name: string,
+    email: string,
+    password: string
+}
