+++ semana19/projeto-labook/src/business/services/idGenerator.ts
@@ -1,0 +1,4 @@
+import { v4 } from "uuid"
+
+
+export const generateId = (): string => v4()
