+++ semana19/projeto-labook/src/index.ts
@@ -6,23 +6,12 @@ import knex from "knex"
 import dotenv from "dotenv"
 import * as jwt from "jsonwebtoken"
 import * as bcrypt from "bcryptjs"
-import { v4 } from "uuid"
+
 import Knex from "knex"
 
 /**************************** CONFIG ******************************/
 
 dotenv.config()
-
-export const connection: Knex = knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      port: 3306,
-      database: process.env.DB_NAME,
-   }
-})
 
 const app: Express = express()
 app.use(express.json())
@@ -30,71 +19,14 @@ app.use(cors())
 
 /**************************** TYPES ******************************/
 
-type AuthenticationData = {
-   id: string
-}
 
-type User = {
-   id: string,
-   name: string,
-   email: string,
-   password: string
-}
 
-enum POST_TYPES {
-   NORMAL = "normal",
-   EVENT = "event"
-}
-
-type Post = {
-   id: string,
-   photo: string,
-   description: string,
-   type: POST_TYPES,
-   createdAt: Date,
-   authorId: string
-}
 
 /**************************** SERVICES ******************************/
 
-const generateId = (): string => v4()
 
-function generateToken(
-   payload: AuthenticationData
-): string {
-   return jwt.sign(
-      payload,
-      process.env.JWT_KEY as string,
-      {
-         expiresIn: process.env.JWT_EXPIRES_IN
-      }
-   )
-}
 
-function getTokenData(
-   token: string
-): AuthenticationData {
-   const result: any = jwt.verify(
-      token,
-      process.env.JWT_KEY as string
-   )
 
-   return { id: result.id, }
-}
-
-const hash = async (
-   plainText: string
-): Promise<string> => {
-   const rounds = Number(process.env.BCRYPT_COST);
-   const salt = await bcrypt.genSalt(rounds);
-   return bcrypt.hash(plainText, salt)
-}
-
-const compare = async (
-   plainText: string, cypherText: string
-): Promise<boolean> => {
-   return bcrypt.compare(plainText, cypherText)
-}
 
 /**************************** ENDPOINTS ******************************/
